extract assets events into AssetsWebhookList, blueprint 'On entry published' instead, update docs
